requirements.txt
openai
SpeechRecognition
PyQt5
python-dotenv
requests
pyaudio
